+++ signin.html
@@ -21,7 +21,7 @@
         <div class="inputModal">
           <label for="email">이메일</label>
           <input type="text" placeholder="이메일" id="email" autoComplete="off" alt="이메일"/>
-          <p id="email-errorMessage" class="warningMessage"></p>
+          <p class="warningMessage email-errorMessage"></p>
         </div>
         <div class="inputModal">
           <label for="password">비밀번호</label>
@@ -29,7 +29,7 @@
           <button class="eyeBtn" type="button">
             <div class="passwordImg"></div>
           </button>
-          <p id="password-errorMessage" class="warningMessage"></p>
+          <p class="warningMessage password-errorMessage"></p>
         </div>
         <button class="form__loginButton" type="submit">
           로그인
@@ -44,6 +44,6 @@
       </div>
     </div>
   </section>
-  <script src="/js/sign.js"></script>
+  <script src="/js/sign.js" type="module"></script>
 </body>
 </html>
